refactor(input): move shared mouse dispatch to src/input/mouse.rs (mode-neutral), dedup ClickTracker
Relocate the shared mouse dispatch (`dispatch_mouse_buttons` /
`dispatch_mouse_wheel`, the pure deciders, geometry, and `MouseInputPlugin`)
out of the library `ozma_terminal` into the host `src/input/mouse.rs`,
keeping it mode-neutral and gated only by `Without<MouseDisabled>`. The
library keeps the apply observer (`on_terminal_mouse_effects`), the
`MouseEffect` / `TerminalMouseEffects` / `TerminalForwardInput` types, and
`OzmaTerminalMouseSet` as the ordering anchor.

- `src/input/bindings.rs`: `OzmaMouseConfig` + `FineModifier` (pub(crate)).
- `src/input/focus.rs`: `MouseDisabled` marker.
- `src/input/mouse.rs`: dispatch + deciders + engine-`Side` geometry; reuses
  `crate::webview_pointer::topmost_surface_at`, `crate::input::gesture::*`,
  `current_terminal_modifiers`, and the host `link_modifier_held`.
- tmux mouse now shares `crate::input::gesture::ClickTracker` (local copy
  deleted from `src/mode/tmux/mouse/decide.rs`).
- `TerminalMouseEffects` gains a `pub fn effects()` accessor so cross-crate
  observers (the moved dispatch tests) can read the decided effects now that
  the field is private.
- Removed the now-dead `link_modifier_held` from `ozma_terminal`'s hyperlink
  module (its only caller, the library dispatch, moved to the host).
