+++ src/ui/copy_mode.rs
@@ -5,13 +5,14 @@
 //! toggle predicate reads `TerminalHandle::selection_type()` to
 //! decide between "start new selection of kind X" and "clear existing".
 
+use crate::input::focus::MouseDisabled;
 use bevy::app::{App, Plugin};
 use bevy::ecs::component::Component;
 use bevy::ecs::entity::Entity;
 use bevy::ecs::event::EntityEvent;
 use bevy::ecs::observer::On;
 use bevy::ecs::system::{Commands, Query};
-use ozma_terminal::{Clipboard, KeyboardDisabled, MouseDisabled};
+use ozma_terminal::{Clipboard, KeyboardDisabled};
 use ozma_tty_engine::{Coalescer, TerminalHandle};
 
 /// Bevy Plugin: registers the two observers and ensures the global
